pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sploitscan"
version = "0.14.0"
description = "SploitScan is a cybersecurity utility designed to provide detailed information on vulnerabilities and associated exploits."
authors = [ { name = "Alexander Hagenah", email = "ah@primepage.de" } ]
license = { file = "LICENSE" }
dynamic = [ "readme" ]
dependencies = [
    "requests==2.32.3",
    "jinja2==3.1.6",
    "openai==1.65.4",
    "google-genai==1.4.0",
    "GitPython==3.1.44",
    "tqdm==4.67.1"
]
requires-python = ">=3.8"

[project.scripts]
sploitscan = "sploitscan.sploitscan:cli"

[project.urls]
homepage = "https://github.com/xaitax/SploitScan"
repository = "https://github.com/xaitax/SploitScan"
documentation = "https://github.com/xaitax/SploitScan#readme"
changelog = "https://github.com/xaitax/SploitScan"

[tool.setuptools]
dynamic.readme = { file = [ "README.md" ], content-type = "text/markdown" }
package-data.sploitscan = [ "templates/*.html", "config.json" ]
